requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==3.2
django-countries==7.2.1
django-jazzmin==2.4.9
Pillow==9.0.0
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
django-ckeditor
